refactor(worker): rename workflow_traits to traits and remove unused adapters
- Rename workflow_traits.rs → traits.rs for brevity
- Remove unused DbEventEmitter (was for V1 workflow)
- Remove unused create_db_agent_loop, create_db_agent_loop_with_registry
- Remove unused create_db_event_emitter, create_openai_provider
- Keep only DbMessageStore (used by activities) and create_db_message_store
- Simplify adapters.rs from ~330 lines to ~170 lines
- Update all imports to use crate::traits instead of crate::workflow_traits
